+++ src/app/[variants]/(main)/knowledge/components/KnowledgeBaseItem/index.tsx
@@ -16,7 +16,6 @@ const useStyles = createStyles(({ css, token, isDarkMode }) => ({
   `,
   container: css`
     cursor: pointer;
-
     padding-block: 4px;
     padding-inline: 8px;
     border-radius: ${token.borderRadius}px;